requirements.txt
streamlit
pytesseract
pillow
python-docx
translate
PyMuPDF
langdetect